+++ src/types/utils.ts
@@ -8,7 +8,7 @@ export type TTreeMapArr = {
     parentId: string;
     id: string;
     lineOrigin?: TBranchLineOrigin;
-}
+};
 
 export type TChildrenMapperReturnEl = TSElement;
 
@@ -21,7 +21,7 @@ export type TTreeClassParams = {
     tree_data: Array<IChartHead>; 
     tsInnerContainer: HTMLElement;
     chartHelper: ChartMainHelper;
-}
+};
 
 export type THeadPointPosition = "top" | "bottom" | "left" | "right";
 
@@ -43,20 +43,51 @@ export type TColorSet = {
 
 export type TTreeToItemHierarchy = Array<{level: number, items: Array<IChartHead>}>;
 
-export type TBranchLineOrigin = "top" | "bottom" | "left" | "right"
+export type TBranchLineOrigin = "top" | "bottom" | "left" | "right";
 
-export type TElementCenterPositions = 'center' | 'top' | 'bottom' | 'right' | 'left'
+export type TElementCenterPositions = 'center' | 'top' | 'bottom' | 'right' | 'left';
 
-export type TLinkerCircleColor = 'brighter' | 'bright500' | 'bright100'
+export type TLinkerCircleColor = 'brighter' | 'bright500' | 'bright100';
 
 export type TLinkerShape = 'symbolAsterisk' | 'symbolCircle' | 'symbolCross' | 
     'symbolDiamond' | 'symbolDiamond2' | 'symbolPlus' | 'symbolSquare' | 
     'symbolSquare2' | 'symbolStar' | 'symbolTriangle' | 'symbolTriangle2' | 
-    'symbolWye' | 'symbolX'
+    'symbolWye' | 'symbolX';
 
 export type THeadImageShape = 'symbolCircle' | 'symbolCross' | 
     'symbolDiamond' | 'symbolDiamond2' | 'symbolSquare' | 
     'symbolSquare2' | 'symbolStar' | 'symbolTriangle' | 'symbolTriangle2' | 
-    'symbolWye'
+    'symbolWye';
 
-export type TEventType = 'chart_head.create' | 'chart_head.expanded' | 'chart_head.collapsed' | 'zooming' | 'library.init'
+export type TEventType = 'chart_head.create' | 'chart_head.expanded' | 'chart_head.collapsed' | 'zooming' | 'library.init';
+
+export type TDropShadow = {
+    x: string;
+    y: string;
+    width: string;
+    height: string;
+    dx: number;
+    dy: number;
+    stdDeviation: number;
+    floodColor: string;
+};
+
+
+export type TColorPallet = {
+    h: number;
+    s: number;
+    l: number;
+    darker: number;
+    brighter: number;
+    bright100: number;
+    dark100: number;
+    gray: number;
+    gray85: number;
+};
+
+export type TTreeType = 'default' | 'cellar' | 'goldenRod' | 'hSpider' | 'hSpiderWalk' | 'radialSpiderLeg' | 'spiderlingsPack' | 'vSpiderWalk';
+
+export type TChartHeadType = 'default' | 'landscape' | 'rounded';
+
+export type TLinkType = 'curveBumpX' | 'curveBumpY' | 'curveBasisClosed' | 'curveLinear' | 
+'curveStep' | 'curveStepAfter' | 'curveStepBefore';